pinus-protobuf: update test.
